@@ -35,14 +35,10 @@
   },
   "jest": {
     "collectCoverage": true,
-    "testPathDirs": [
+    "roots": [
       "./__tests__"
     ],
-    "scriptPreprocessor": "./node_modules/babel-jest",
-    "testFileExtensions": [
-      "jsx",
-      "js"
-    ],
+    "transform": {".*": "./node_modules/babel-jest"},
     "moduleFileExtensions": [
       "jsx",
       "js"
@@ -68,18 +64,21 @@
     "babelify": "^7.2.0",
     "browserify": "^12.0.1",
     "extensify": "^1.0.2",
-    "jest": "^0.1.40",
+    "jest": "^23.6.0",
     "node-sass": "^3.4.2",
     "nodemon": "^1.8.1",
     "postcss-cli": "^4.1.0",
-    "react-addons-test-utils": "^15.4.0",
+    "react-addons-test-utils": "^15.6.2",
     "reactify": "^1.1.1",
     "watchify": "^3.7.0"
   },
   "dependencies": {
-    "classnames": "^2.2.3",
+    "chrono-node": "^1.3.5",
+    "classnames": "^2.2.6",
+    "jest-cli": "^23.6.0",
     "moment": "^2.11.0",
-    "react": "^15.4.0",
-    "react-dom": "^15.4.0"
+    "prop-types": "^15.6.2",
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0"
   }
 }
