@@ -1,16 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
 import moment from 'moment';
 import WeekRow from './week-row';
 
 class MonthView extends React.Component {
 
-    static propTypes = { 
-        date: React.PropTypes.instanceOf(moment).isRequired,
-        handleSelection: React.PropTypes.func.isRequired,
-        minDate: React.PropTypes.instanceOf(moment),
-        maxDate: React.PropTypes.instanceOf(moment),
-        selectedDate: React.PropTypes.instanceOf(moment)
+    static propTypes = {
+        date: PropTypes.instanceOf(moment).isRequired,
+        handleSelection: PropTypes.func.isRequired,
+        minDate: PropTypes.instanceOf(moment),
+        maxDate: PropTypes.instanceOf(moment),
+        selectedDate: PropTypes.instanceOf(moment)
     };
 
     constructor(props) {
@@ -39,14 +40,14 @@ class MonthView extends React.Component {
             if(dayOfWeek === 0 || i === 0) {
                  weeks.push(<WeekRow selectedDate={this.props.selectedDate} maxDate={this.props.maxDate} minDate={this.props.minDate} handleSelection={this.props.handleSelection} month={currentMonth} date={dateToPass} key={currentDate.date()} />);
             }
-           
+
 
             currentDate.add(1, "days");
             i++;
         }
 
         return weeks;
-        
+
     }
 
     render() {
@@ -57,5 +58,6 @@ class MonthView extends React.Component {
         );
     }
 }
+
 
 module.exports = MonthView;
